+++ app/src/main/java/com/drawiin/mymovielist/core/arch/OnStartEffect.kt
@@ -1,0 +1,2 @@
+package com.drawiin.mymovielist.core.arch
+