+++ app/src/main/java/com/drawiin/mymovielist/core/uikit/foundation/Type.kt
@@ -1,0 +1,5 @@
+package com.drawiin.mymovielist.core.uikit.foundation
+
+import androidx.compose.material3.Typography
+
+val AppTypography = Typography()