+++ app/src/test/java/com/drawiin/mymovielist/core/test/InstantDispatcherRule.kt
@@ -1,0 +1,30 @@
+package com.drawiin.mymovielist.core.test
+
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.ExperimentalCoroutinesApi
+import kotlinx.coroutines.test.StandardTestDispatcher
+import kotlinx.coroutines.test.resetMain
+import kotlinx.coroutines.test.setMain
+import org.junit.rules.TestRule
+import org.junit.runner.Description
+import org.junit.runners.model.Statement
+
+@OptIn(ExperimentalCoroutinesApi::class)
+class InstantDispatcherRule: TestRule {
+    private val testDispatcher = StandardTestDispatcher()
+
+    override fun apply(base: Statement, description: Description): Statement {
+        return object : Statement() {
+            override fun evaluate() {
+                Dispatchers.setMain(testDispatcher)
+                try {
+                    base.evaluate()
+                } finally {
+                    Dispatchers.resetMain()
+                }
+            }
+        }
+    }
+
+    val dispatcher get() = testDispatcher
+}